refactor(workflow): split workflow runtime into modules
